requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
stripe==10.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
